+++ spec/integrations/pro/routing/share_groups/warmup_with_share_group_spec.rb
@@ -44,7 +44,7 @@ draw_routes(create_topics: false) do
 
   share_group "sg" do
     topic "share-topic" do
-      consumer Class.new(Karafka::BaseConsumer)
+      consumer Class.new(Karafka::ShareConsumer)
     end
   end
 end