+++ spec/integrations/routing/limited_scope/with_wildcard_topics_char_class_spec.rb
@@ -7,15 +7,15 @@ setup_karafka
 
 draw_routes(create_topics: false) do
   topic "topic-1" do
-    consumer Class.new
+    consumer Class.new(Karafka::BaseConsumer)
   end
 
   topic "topic-2" do
-    consumer Class.new
+    consumer Class.new(Karafka::BaseConsumer)
   end
 
   topic "topic-3" do
-    consumer Class.new
+    consumer Class.new(Karafka::BaseConsumer)
   end
 end