+++ spec/lib/karafka/pro/cli/contracts/server_spec.rb
@@ -83,7 +83,7 @@ RSpec.describe_current do
     context "when we have pattern matching defined" do
       before do
         Karafka::App.routes.pattern(/test/) do
-          consumer Class.new
+          consumer Class.new(Karafka::BaseConsumer)
         end
       end
 
@@ -105,7 +105,7 @@ RSpec.describe_current do
     context "when we have pattern matching defined" do
       before do
         Karafka::App.routes.pattern(/test/) do
-          consumer Class.new
+          consumer Class.new(Karafka::BaseConsumer)
         end
       end